Removing duplicate code from assembly matrix
